+++ app/src/main/java/com/lambdaschool/sprint2_challenge/Controller/MainActivity.kt
@@ -1,0 +1,13 @@
+package com.lambdaschool.sprint2_challenge.Controller
+
+import android.support.v7.app.AppCompatActivity
+import android.os.Bundle
+import com.lambdaschool.sprint2_challenge.R
+
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+}
